LayoutTests/inspector/canvas/context-attributes-expected.txt

Test that CanvasAgent is able to send context attributes.


== Running test suite: Canvas.contextAttributes
-- Running test case: Create2DCanvasContext
Added canvas.
PASS: Canvas context should be "2D".
{}

-- Running test case: CreateBitmapRendererCanvasContext
Added canvas.
PASS: Canvas context should be "Bitmap Renderer".
{
  "alpha": true
}

-- Running test case: CreateBitmapRendererCanvasContextWithAttributes
Added canvas.
PASS: Canvas context should be "Bitmap Renderer".
{
  "alpha": false
}
PASS: Canvas context should have attribute "alpha" with value "false".

-- Running test case: CreateWebGLCanvasContext
Added canvas.
PASS: Canvas context should be "WebGL".
{
  "alpha": true,
  "depth": true,
  "stencil": false,
  "antialias": true,
  "premultipliedAlpha": true,
  "preserveDrawingBuffer": false,
  "failIfMajorPerformanceCaveat": false
}

-- Running test case: CreateWebGLCanvasContextWithAttributes
Added canvas.
PASS: Canvas context should be "WebGL".
{
  "alpha": false,
  "depth": true,
  "stencil": false,
  "antialias": true,
  "premultipliedAlpha": true,
  "preserveDrawingBuffer": false,
  "failIfMajorPerformanceCaveat": false
}
PASS: Canvas context should have attribute "alpha" with value "false".
